webapi/webapi-webspeech-w3c-tests/webspeech/blink/speechrecognitionerror-basics.html

<!--
Test convert from Blink:
https://github.com/crosswalk-project/blink-crosswalk/blob/master/LayoutTests/fast/speech/scripted/speechrecognitionerror-basics.html
-->

<!DOCTYPE html>
<meta charset="utf-8">
<title>Web Speech API Test: the basics of the SpeechRecognitionError interface</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://dvcs.w3.org/hg/speech-api/raw-file/tip/speechapi.html">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

test(function () {
  assert_true('webkitSpeechRecognitionError' in window);
}, "Test SpeechRecognitionError existence");

test(function () {
  assert_not_equals(webkitSpeechRecognitionError, null);
}, "Test SpeechRecognitionError is not null");

</script>
